rename temp-name parent folder to dialog
