+++ src/components/component-names.ts
@@ -36,7 +36,7 @@ export const LIMIT_TEXT_NAME: string = 'm-limit-text';
 export const LINK_NAME: string = 'm-link';
 export const LIST_ITEM_NAME: string = 'm-list-item';
 export const MESSAGE_NAME: string = 'm-message';
-export const TEMP_NAME: string = 'm-temp-name';
+export const TEMP_NAME: string = 'm-dialog';
 export const NAVBAR_NAME: string = 'm-navbar';
 export const NAVBAR_ITEM_NAME: string = 'm-navbar-item';
 export const MENU_NAME: string = 'm-menu';
