+++ src/components/phone-number/phone-number.spec.ignore.ts
@@ -331,7 +331,7 @@
 //             Vue.nextTick(() => {
 //                 (vm as any).visible = false;
 //                 Vue.nextTick(() => {
-//                     (vm as any).temp-name = true;
+//                     (vm as any).dialog = true;
 //                     Vue.nextTick(() => {
 //                         (vm as any).visible = true;
 //                         Vue.nextTick(() => {
