+++ src/utils/vue/vue.ts
@@ -7,8 +7,8 @@ import { Messages } from '../i18n/i18n';
 import { Licenses } from '../license/license';
 import { Logger } from '../logger/logger';
 import { MediaQueries } from '../media-queries/media-queries';
-import { AlertFunction } from '../temp-name/alert';
-import { ConfirmFunction } from '../temp-name/confirm';
+import { AlertFunction } from '../dialog/alert';
+import { ConfirmFunction } from '../dialog/confirm';
 import { Modul } from '../modul/modul';
 
 // TODO: explore usage of TS declare syntax
